// src/app/modules/master/models/formula-list.ts
/*
 * Copyright (c) 2021 All Rights Reserved.
 *
 * Unauthorized copying of this file, via any medium is strictly prohibited.
 * Proprietary and confidential.
 *
 * Written by:
 *   - Satrya Wiguna <satrya@freshcms.net>
 */

export type FormulaList = {
  formula_name: string;
  formula_category: 'P'; // P: Payroll
  description: string;
};